requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
deepdiff==8.4.2
Django==5.1.7
django-mail-templated==2.6.5
django-simple-history==3.8.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.1.1
ordered-set==4.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==78.1.0
sqlparse==0.5.3
urllib3==2.3.0
wheel==0.45.1